requirements.txt
# requirements.txt (업데이트 버전)
langchain>=0.3.5
langchain-openai>=0.2.1
langchain-community>=0.3.5
langchain-chroma>=0.1.1
chromadb>=0.5.5
rank-bm25>=0.2.2

python-dotenv
pandas
pdfplumber